errors module
